+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.2 - April 6, 2025
+* Add App Category on App Usage
+* To get category better to use minimum Android 26
+
 ## 0.4.1 - April 6, 2025
 * Improvement fetch app installed on android 13